requirements.txt
fastapi
uvicorn[standard]
pypdf
sentence-transformers
numpy
torch
faiss-cpu
requests
python-multipart